Linux: Creates a Types.h header that matches between architectures
